feat: restaurants api call
